+++ tests/test_project_submission.py
@@ -9,11 +9,8 @@ from synapseclient.annotations import to_submission_status_annotations
 
 import challengeutils.utils
 import challengeutils.project_submission
-#from challengeutils.project_submission import _archive_project_submission
-#from challengeutils.project_submission import archive_project_submission
-#from challengeutils.project_submission import archive_project_submissions
 from challengeutils.project_submission import join_evaluations
-#from challengeutils.project_submission import archive_and_attach_project_submissions
+
 
 
 SYN = mock.create_autospec(synapseclient.Synapse)
@@ -191,37 +188,6 @@ SUBMISSION_QUEUEID = '3'
 #         patch_query.assert_called_once_with(SYN, query)
 #         assert patch_archive.call_count == 2
 #         assert archived == ['syn1234', 'syn1234']
-
-
-def test_join_queues():
-    """Archive and attach writeups"""
-    writeups = [{"submitterId": '123',
-                 'objectId': '2222'},
-                {"submitterId": '234',
-                 'objectId': '5555'}]
-
-    submissions = [{"submitterId": '123',
-                    'objectId': '3333'},
-                   {"submitterId": '234',
-                    'objectId': '4444'}]
-
-    firstquery = "select * from evaluation_2"
-    secondquery = "select * from evaluation_3"
-    expected_dict = [{'submitterId': '123',
-                      'objectId_x': '2222',
-                      'objectId_y': '3333'},
-                     {'submitterId': '234',
-                      'objectId_x': '5555',
-                      'objectId_y': '4444'}]
-    expecteddf = pd.DataFrame(expected_dict)
-    with patch.object(challengeutils.utils,
-                      "evaluation_queue_query",
-                      side_effect=[writeups,
-                                   submissions]) as patch_query:
-        joineddf = join_evaluations(SYN, '2', '3', "submitterId")
-        calls = [mock.call(SYN, firstquery), mock.call(SYN, secondquery)]
-        patch_query.assert_has_calls(calls)
-        assert joineddf.equals(expecteddf[joineddf.columns])
 
 
 # def test_archive_and_attach_project_submissions():
